requirements.txt
torch>=2.1.0
torchvision>=0.16.0
ultralytics>=8.0.0
opencv-python>=4.8.0
pillow>=10.0.0
scikit-image>=0.21.0
scikit-learn>=1.3.0
streamlit>=1.28.0
streamlit-drawable-canvas>=0.9.0
numpy>=1.26.0
pandas>=2.1.0
matplotlib>=3.7.0
plotly>=5.9.0
seaborn>=0.13.0
tqdm>=4.66.0